style: add missing Sun style Javadoc and parameters
