+++ android/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:label="@string/app_name"
         tools:ignore="UnusedAttribute">
         <activity
-            android:name="tdt4240.group2.AndroidLauncher"
+            android:name="prog.sumo.AndroidLauncher"
             android:label="@string/app_name"
             android:screenOrientation="landscape"
             android:configChanges="keyboard|keyboardHidden|navigation|orientation|screenSize|screenLayout"
