+++ android/src/prog/sumo/AndroidFirebaseScoreAPI.java
@@ -1,6 +1,5 @@
-package tdt4240.group2;
+package prog.sumo;
 
-import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
 import java.util.Map;
