+++ core/src/prog/sumo/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Sumo Battle Tap is a game.
+ */
+package prog.sumo;
