--- core/src/tdt4240/group2/ScoreAPI.java
@@ -1,9 +1,0 @@
-package tdt4240.group2;
-
-import java.util.Map;
-
-public interface ScoreAPI {
-    void incrementScore(String characterName);
-
-    void getScores(Map<String, Long> scores);
-}
