--- core/src/tdt4240/group2/SumoBattleTapGame.java
@@ -1,36 +1,0 @@
-package tdt4240.group2;
-
-import com.badlogic.gdx.ApplicationAdapter;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.utils.ScreenUtils;
-
-public class SumoBattleTapGame extends ApplicationAdapter {
-    private final ScoreAPI scoreAPI;
-    SpriteBatch batch;
-    Texture img;
-
-    public SumoBattleTapGame(ScoreAPI scoreAPI) {
-        this.scoreAPI = scoreAPI;
-    }
-
-    @Override
-    public void create() {
-        batch = new SpriteBatch();
-        img = new Texture("badlogic.jpg");
-    }
-
-    @Override
-    public void render() {
-        ScreenUtils.clear(1, 0, 0, 1);
-        batch.begin();
-        batch.draw(img, 0, 0);
-        batch.end();
-    }
-
-    @Override
-    public void dispose() {
-        batch.dispose();
-        img.dispose();
-    }
-}
