--- desktop/src/tdt4240/group2/DesktopLauncher.java
@@ -1,18 +1,0 @@
-package tdt4240.group2;
-
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-
-import tdt4240.group2.SumoBattleTapGame;
-
-// Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
-public class DesktopLauncher {
-    public static void main(String[] arg) {
-        Lwjgl3ApplicationConfiguration config =
-                new Lwjgl3ApplicationConfiguration();
-        config.setForegroundFPS(60);
-        config.setTitle("Sumo Battle Tap");
-        new Lwjgl3Application(new SumoBattleTapGame(new DesktopScoreAPI()),
-                config);
-    }
-}
