--- desktop/src/tdt4240/group2/DesktopScoreAPI.java
@@ -1,15 +1,0 @@
-package tdt4240.group2;
-
-import java.util.Map;
-
-public class DesktopScoreAPI implements ScoreAPI {
-    @Override
-    public void incrementScore(String characterName) {
-        System.err.println("Score API not implemented in desktop version!");
-    }
-
-    @Override
-    public void getScores(Map<String, Long> scores) {
-        System.err.println("Score API not implemented in desktop version!");
-    }
-}
